+++ python/test/unit/fem/test_nonlinear_assembler.py
@@ -57,7 +57,7 @@ def test_matrix_assembly_block():
         return numpy.cos(x[0]) * numpy.cos(x[1])
 
     facetdim = mesh.topology.dim - 1
-    mf = dolfinx.MeshFunction("size_t", mesh, facetdim, 0)
+    mf = dolfinx.MeshFunction("int", mesh, facetdim, 0)
     mf.mark(boundary, 1)
     bndry_facets = numpy.where(mf.values == 1)[0]
 
@@ -266,7 +266,7 @@ def test_assembly_solve_block():
         return numpy.logical_or(x[0] < 1.0e-6, x[0] > 1.0 - 1.0e-6)
 
     facetdim = mesh.topology.dim - 1
-    mf = dolfinx.MeshFunction("size_t", mesh, facetdim, 0)
+    mf = dolfinx.MeshFunction("int", mesh, facetdim, 0)
     mf.mark(boundary, 1)
     bndry_facets = numpy.where(mf.values == 1)[0]
 
@@ -472,7 +472,7 @@ def test_assembly_solve_taylor_hood(mesh):
     u_bc_1.interpolate(lambda x: numpy.row_stack(tuple(numpy.sin(x[j]) for j in range(gdim))))
 
     facetdim = mesh.topology.dim - 1
-    mf = dolfinx.MeshFunction("size_t", mesh, facetdim, 0)
+    mf = dolfinx.MeshFunction("int", mesh, facetdim, 0)
     mf.mark(boundary0, 1)
     mf.mark(boundary1, 2)
     bndry_facets0 = numpy.where(mf.values == 1)[0]
